Created structure of package, added notebooks, finaliyed preprocessing.py
